MSVC 7.0 (Visual C/C++ 2002) support, and touhou6 preset (#839)
* Introduce MSVC7.0 (Visual Studio 2002) toolchain

This renames the msvc7.0 toolchain to msvc7.1 (as it is based on Visual
Studio 2003, which is msvc7.1), and introduces a new msvc7.0 toolchain
based on Visual Studio 2002.

* Add presets for Touhou 6 decompilation project

* Remove noexec from /sandbox/tmp in docker

The MSVC toolchains store intermediate artifacts in the /sandbox/tmp
folder, and need to be able to mmap them with PROT_EXEC rights, which is
prevented by the noexec flag that docker automatically puts on all
tmpfs. To avoid this, we need to ask docker to remove that flag.