Update Frogger's compiler preset. (#898)
* Update Frogger's compiler preset.

More investigation was done, and it was discovered that there was an official release of PsyQ GCC 2.6.3. However, it has not surfaced online. I have compiled a Win32 version which is compatible with the original PsyQ toolchain, but this is likely unnecessary for decomp.me

* Changed my mind about adding -G8.

The flag -G8 is correct when building the full game, but it produces less accurate code on decomp.me for some reason.