Toggle problems by clicking anywhere on the bar (#1150)
Problems bar can now be toggled anywhere on the bar

Previously the problems bar could only be toggled by clicking the small icon on the very far right. It would be more pleasant of a user experience to be able to simply click the bar anywhere to toggle it, as it is much quicker and requires less precision.