Fix indentation when braces are on new lines (#1164)
* Roll our own cpp LRLanguage

This fixes a bug when using a coding style with braces on a new line. Previously, it was inconsistent whether braces were indented (GNU style) or not (Allman style). This commit standardizes on Allman style.

* Fix try-catch and do-while recognition

Both sections are needed; `indentOnInput` is required to call the indent checker, and `DoStatement` is required to provide the correct indent to the indent checker.

* Add copyright line from upstream because I am forgetful

* style changes

---------

Co-authored-by: Ethan <ethteck@gmail.com>