Update default font & increase size
This pulls in JetBrains Mono as a variable web font (2 KiB)
and uses it as the default font family. The default size is
updated from 11px to 13px to improve legibility.

Additionally, an appearance setting "Enable ligatures" is
added, allowing users to disable the font ligatures that
are included in JetBrains Mono (or their own configured
font).